backend/pyproject.toml
[project]
name = "django-channel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-phonenumber-field>=8.4.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "phonenumbers>=9.0.27",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.9",
    "pytest-asyncio>=0.24",
    "ruff>=0.11",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[tool.ruff]
line-length = 120
exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
